Moved models and routes to folders
