+++ README.md
@@ -4,7 +4,8 @@
 
 1. Open a terminal and go to the server folder. Make sure you have **pipenv** installed (`pip install pipenv`)
 2. Install the dependencies with `pipenv install`. This also createa a virtual environment, if there isn't one already
-3. Activate the virtual environment and start the app with `pipenv run flask run`
+3. Activate the virtual environment with `pipenv shell`
+4. Start the app with `flask run`
 
 ## Adding Postgres database:
