+++ server/models/__init__.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+from marshmallow import Schema
+
+# initialize our db
+db = SQLAlchemy()
+
+from .User import Customer, CustomerSchema
